% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Get datasets by ID.}
\usage{
datasets(ids, verbose = FALSE)
}
\arguments{
\item{ids}{Dataset IDs.}

\item{verbose}{Verbose.}
}
\value{
List of datasets.
}
\description{
Get datasets by ID.
}

