% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasetsSpecialCollection}
\alias{datasetsSpecialCollection}
\title{Get datasets by special collection.}
\usage{
datasetsSpecialCollection(id, verbose = FALSE)
}
\arguments{
\item{id}{Special collection id.}

\item{verbose}{Verbose.}
}
\value{
Datasets data frame.
}
\description{
Get datasets by special collection.
}
