% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatdatasetrecord.R
\name{flatdatasetrecord}
\alias{flatdatasetrecord}
\title{Get a flat table of an IMIS datasets record using DasIDs}
\usage{
flatdatasetrecord(dasid)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This generates a flat table of an IMIS dataset record using DasIDs
}
\examples{
dasrecord <- flatdatasetrecord(dasid = 4662)
dasrecord <- flatdatasetrecord(dasid = c("1884","618", "5780" ))
}
