% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdascitations.r
\name{getdascitations}
\alias{getdascitations}
\title{Get citations and licence from IMIS using the output of the datasets function}
\usage{
getdascitations(datasetrecords)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This function gets the citations and DOIs from IMIS using the output of the datasets function
}
\examples{
meta <- getdascitations(datasets(4662))
meta <- getdascitations(datasets(c("1884","618", "5780" )))
}
