% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdascontactpersons.R
\name{getdascontacts}
\alias{getdascontacts}
\title{Get the contacts from IMIS using the output of the datasets function}
\usage{
getdascontacts(datasetrecords)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This function gets the themes associated to datasets using the output of the datasets function
}
\examples{
dasthemes <- getdascontacts(datasets(4662))
dasthemes <- getdascontacts(datasets(c("1884","618", "5780" )))
}
