% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getkeywords.R
\name{getdaskeywords}
\alias{getdaskeywords}
\title{Get the keywords from IMIS using the output of the datasets function}
\usage{
getdaskeywords(datasetrecords)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This function gets the keywords associated to datasets using the output of the datasets function
}
\examples{
dasthemes <- getdaskeywords(datasets(4662))
dasthemes <- getdaskeywords(datasets(c("1884","618", "5780" )))
}
