% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdatasetthemes.R
\name{getdasthemes}
\alias{getdasthemes}
\title{Get the themes from IMIS based on dasids}
\usage{
getdasthemes(dasid)
}
\arguments{
\item{dasid}{mandatory parameter, the dasid of the dataset}
}
\description{
This function gets the themes associated to  dasids
}
\examples{
dasthemes <- getdasthemes(4662)
dasthemes <- getdasthemes(dasid = c("1884","618", "5780" ))
}
