% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdasurls.R
\name{getdasurls}
\alias{getdasurls}
\title{Get the urls from IMIS using the output of the datasets function}
\usage{
getdasurls(datasetrecords)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This function gets the themes associated to datasets using the output of the datasets function
}
\examples{
dasurls <- dataseturls(datasets(4662))
dasurls <- dataseturls(datasets(c("1884","618", "5780" )))
}
