% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdatasetrecords.R
\name{qcdatasetrecords}
\alias{qcdatasetrecords}
\title{Get a QC value for datasets}
\usage{
qcdatasetrecords(flatdatasetrecords)
}
\arguments{
\item{flatdatasetrecords}{mandatory parameter, the output of the flatdatasetrecords function}
}
\description{
This generates a flat table of an IMIS dataset record using DasIDs
}
\examples{
QCedrecords <- qcdatasetrecords(flatdatasetrecords(dasid = 4662))
QCedrecords <- qcdatasetrecords(flatdatasetrecords = c("1884","618", "5780" ))
}
