% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getkeywords.R
\name{getdaskeywords}
\alias{getdaskeywords}
\title{Get the keywords from IMIS based on dasids}
\usage{
getdaskeywords(dasid)
}
\arguments{
\item{dasid}{mandatory parameter, the dasid of the dataset}
}
\description{
This function gets the keywords associated to  dasids
}
\examples{
dasthemes <- getdaskeywords(4662)
dasthemes <- getdaskeywords(dasid = c("1884","618", "5780" ))
}
