% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdatasetthemes.R
\name{getdasthemes}
\alias{getdasthemes}
\title{Get the themes from IMIS using the output of the datasets function}
\usage{
getdasthemes(datasetrecords)
}
\arguments{
\item{datasetrecords}{mandatory parameter, the output of the datasets function}
}
\description{
This function gets the themes associated to datasets using the output of the datasets function
}
\examples{
dasthemes <- getdasthemes(datasets(4662))
dasthemes <- getdasthemes(datasets(c("1884","618", "5780" )))
}
